+++ frontend/src/services/routines.tsx
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { API_ENDPOINTS } from "../config";
+import { API_ENDPOINTS } from "@/config";
 import { Routine } from "@/types/routine_types";
 
 export const fetchRoutines = async (): Promise<Routine[]> => {
@@ -27,21 +27,36 @@ export const fetchRoutine = async (routineId: string): Promise<Routine> => {
 };
 
 export const createRoutine = async (data: any) => {
-    await axios.post(
-        `${API_ENDPOINTS.BASE_URL}${API_ENDPOINTS.ROUTINES}`,
-        data
-    );
+    try {
+        await axios.post(
+            `${API_ENDPOINTS.BASE_URL}${API_ENDPOINTS.ROUTINES}`,
+            data
+        );
+    } catch (error) {
+        console.error("Error while creating routine:", error);
+        throw error;
+    }
 };
 
 export const updateRoutine = async (id: string, data: any) => {
-    await axios.put(
-        `${API_ENDPOINTS.BASE_URL}${API_ENDPOINTS.ROUTINES}/${id}`,
-        data
-    );
+    try {
+        await axios.put(
+            `${API_ENDPOINTS.BASE_URL}${API_ENDPOINTS.ROUTINES}/${id}`,
+            data
+        );
+    } catch (error) {
+        console.error("Error while updating routine:", error);
+        throw error;
+    }
 };
 
 export const deleteRoutine = async (id: string) => {
-    await axios.delete(
-        `${API_ENDPOINTS.BASE_URL}${API_ENDPOINTS.ROUTINES}/${id}`
-    );
+    try {
+        await axios.delete(
+            `${API_ENDPOINTS.BASE_URL}${API_ENDPOINTS.ROUTINES}/${id}`
+        );
+    } catch (error) {
+        console.error("Error while deleting routine:", error);
+        throw error;
+    }
 };
